pull

+++ app/controllers/Admin.php
@@ -1,7 +1,14 @@
 <?php
-  class Admin extends Controller {
-    public function __construct(){
 
+use PHPMailer\PHPMailer\PHPMailer;
+use PHPMailer\PHPMailer\SMTP;
+use PHPMailer\PHPMailer\Exception;
+  class Admin extends Controller {
+    
+    private $mail;
+    
+    public function __construct(){
+ 
       $this->adminModel=$this->model('Admins');
       $this->userModel=$this->model('User');
       $this->creditModel=$this->model('Credit_amount');
@@ -26,6 +33,25 @@
       $this->garbage_Model=$this->model('Garbage_Stock');
       $this->MailSubscriptionModel = $this->model('Mail_Subscriptions');
       $this->Customer_Credit_Model = $this->model('Customer_Credit');
+
+       //  $this->mail = new PHPMailer();
+          //  $this->mail->isSMTP();
+          //  $this->mail->Host = 'smtp.gmail.com';
+          //  $this->mail->Port = 587;
+          //  $this->mail->Username = 'ecoplusgroupproject@gmail.com';
+          //  $this->mail->Password = 'zzruvawrzshhafbk';
+          //  $this->mail->SMTPSecure = 'tls';
+          //  $this->mail->SMTPAuth = true;
+           
+
+              // Setup PHPMailer
+        $this->mail = new PHPMailer();
+        $this->mail->isSMTP();
+        $this->mail->Host = 'sandbox.smtp.mailtrap.io';
+        $this->mail->SMTPAuth = true;
+        $this->mail->Port = 2525;
+        $this->mail->Username = 'f4ab65cd067d1f';
+        $this->mail->Password = '111c78b575960b';
 
 
       if(!isLoggedIn('admin_id')  && !isLoggedIn('superadmin_id')){
@@ -147,7 +173,12 @@
       $cm_id = $this->center_managerModel->getCenterManagerByID($id);
       $this->center_managerModel->delete_centermanager($id);
       $center_managers = $this->center_managerModel->get_center_managers();
-      deleteImage("C:\\xampp\\htdocs\\ecoplus\\public\\img\\img_upload\\center_manager\\" . $cm_id->image);
+      if($cm_id->image=="profile.png"){
+
+      }else{
+        deleteImage("C:\\xampp\\htdocs\\ecoplus\\public\\img\\img_upload\\center_manager\\" . $cm_id->image);
+
+      }
       $data = [
         'center_managers' => $center_managers,
         'confirm_delete' =>'',
@@ -255,12 +286,20 @@
 
 
         // Validate Password
-        if(empty($data['password'])){
+        if (empty($data['password'])) {
           $data['password_err'] = 'Please enter password';
-        } elseif(strlen($data['password']) < 6){
-          $data['password_err'] = 'Password must be at least 6 characters';
-        }
-
+      } elseif (strlen($data['password']) < 8 || strlen($data['password']) > 30) {
+          $data['password_err'] = 'Password must be between 8 and 30 characters';
+      } elseif (!preg_match('/[^\w\s]/', $data['password'])) {
+          $data['password_err'] = 'Password must include at least one symbol';
+      } elseif (!preg_match('/[A-Z]/', $data['password'])) {
+          $data['password_err'] = 'Password must include at least one uppercase letter';
+      } elseif (!preg_match('/[a-z]/', $data['password'])) {
+          $data['password_err'] = 'Password must include at least one lowercase letter';
+      } elseif (!preg_match('/[0-9]/', $data['password'])) {
+          $data['password_err'] = 'Password must include at least one number';
+      }
+      
         // Validate Confirm Password
         if(empty($data['confirm_password'])){
           $data['confirm_password_err'] = 'Please confirm password';
@@ -274,14 +313,68 @@
         
 
         if(empty($data['email_err']) && empty($data['name_err']) && empty($data['password_err']) && empty($data['confirm_password_err']) && empty($data['contact_no_err']) && empty($data['nic_err']) && empty($data['address_err']) && empty($data['dob_err'])){
-          // Validated
+          // // Validated
+          // $data['password'] = password_hash($data['password'], PASSWORD_DEFAULT);
+          // if($this->center_managerModel->register_center_manager($data)){
+          //   $data['registered']='True';        
+          //   $this->view('admin/center_managers_add',$data);
+          // } else {
+          //   die('Something went wrong');
+          // }
+
+
+          $pw=$data['password'];
           $data['password'] = password_hash($data['password'], PASSWORD_DEFAULT);
-          if($this->center_managerModel->register_center_manager($data)){
-            $data['registered']='True';        
-            $this->view('admin/center_managers_add',$data);
-          } else {
-            die('Something went wrong');
+          $selector = bin2hex(random_bytes(8));
+          $token = random_bytes(32);
+         
+          $url = 'http://localhost/ecoplus/users/register_success_CMAdmin?selector='.$selector.'&validator='.bin2hex($token).'&email='.urlencode($data['email']);            
+          //Expiration date will last for half an hour
+          $expires = date("U") + 1800;
+          if(!$this->userModel->deleteEmailCM_Admin($data['email'])){
+            header("Location: " . URLROOT . "");        
+
           }
+         
+          
+          $hashedToken = password_hash($token, PASSWORD_DEFAULT);
+          $data['selector']=$selector;
+          $data['hashedToken']=$hashedToken;
+          $data['expires']=$expires;
+          // Register User
+          if($this->userModel->register_confirm_cm_admin($data)){
+            
+            $usersEmail = $data['email'];
+
+      
+            $subject = "Login to your account";
+            $message = "<p>We recieved a login request.</p>";
+            $message .= "<p>Here is your login link: </p>";
+            $message .= "<a href='".$url."'>".$url."</a>";
+
+            $this->mail->setFrom('ecoplusgroupproject@gmail.com');
+            $this->mail->isHTML(true);
+            $this->mail->Subject = $subject;
+            $this->mail->Body = $message;
+            $this->mail->addAddress( $usersEmail);
+
+            if (!$this->mail->send()) {
+              header("Location: " . URLROOT . "");        
+          } else {
+              $data['registered'] = 'True';        
+              $this->view('admin/center_managers_add', $data);
+          }
+          
+
+          
+          }
+           else {
+            redirect('users/login');
+          }
+
+
+
+          
         }
         else{
          
@@ -746,6 +839,10 @@
       $no_of_collectors = $this->collector_model->get_no_of_Collectors($center_id);
       $no_of_workers = $this->center_workers_model->get_no_of_center_workers($center_id);
       $total_requests = $this->requests_model->get_total_requests_by_region($region);
+      $total_garbage = $this->garbage_Model->get_total_garbage_by_centerId($center_id);
+      $marked_holidays = $this->center_managerModel->get_marked_holidays($region);
+
+      //var_dump($total_garbage);
       
       if($_SERVER['REQUEST_METHOD'] == 'POST'){  
         $data = [
@@ -756,9 +853,11 @@
           'no_of_workers'=>$no_of_workers,
           'center_manager' =>$center_manager,
           'total_requests'=>$total_requests,
+          'total_garbage' => $total_garbage->total_garbage,
           'lattitude'=>trim($_POST['latittude']),
           'longitude'=>trim($_POST['longitude']),
           'radius'=>trim($_POST['radius']),
+          'marked_holidays'=> $marked_holidays,
           'center_id'=>$center_id,
           'region'=> $region
       ];
@@ -779,11 +878,13 @@
         'no_of_workers'=>$no_of_workers,
         'center_manager' =>$center_manager,
         'total_requests'=>$total_requests,
+        'total_garbage' => $total_garbage->total_garbage,
+        'marked_holidays'=> $marked_holidays,
         'lattitude'=>'',
         'longitude'=>'',
         'radius'=>'',
         'center_id'=>$center_id,
-          'region'=> $region
+        'region'=> $region
       ];
       
       $this->view('admin/center_main', $data);        
@@ -795,6 +896,7 @@
        $no_of_collectors = $this->collector_model->get_no_of_Collectors($center_id);
        $total_requests = $this->requests_model->get_total_requests_by_region($center->region);
        $no_of_workers = $this->center_workers_model->get_no_of_center_workers($center_id);
+       $total_garbage = $this->garbage_Model->get_total_garbage_by_centerId($center_id);
 
       if($_SERVER['REQUEST_METHOD'] == 'POST'){
        
@@ -828,7 +930,9 @@
             'center_manager'=>trim($_POST['centerManager']),
             'no_of_collectors' =>$no_of_collectors,
             'no_of_workers'=>$no_of_workers,
-            'total_requests'=>$total_requests
+            'total_requests'=>$total_requests,
+            'total_garbage' => $total_garbage->total_garbage,
+
           ];
         
           header("Location: " . URLROOT . "/admin/center_main/$center_id/$center->region");        
@@ -849,7 +953,8 @@
           'no_of_collectors' =>$no_of_collectors,
           'no_of_workers'=>$no_of_workers,
           'center_manager' =>$center_manager,
-          'total_requests'=>$total_requests
+          'total_requests'=>$total_requests,
+          'total_garbage' => $total_garbage->total_garbage
         ];
         $this->view('admin/center_main', $data);
       }
@@ -1392,9 +1497,8 @@
         $_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_STRING);
 
         $data=[ 'name' => trim($_POST['name']),
-        'profile' => $_FILES['profile_image'],
          'contact_no' => trim($_POST['contact_no']),
-        'profile_image_name' => trim($_POST['email']).'_'.$_FILES['profile_image']['name'],
+        'profile_image_name' => "profile.png",
         'nic' => trim($_POST['nic']),
         'address' => trim($_POST['address']),
         'dob' => trim($_POST['dob']),
@@ -1476,25 +1580,10 @@
             $data['confirm_password_err'] = 'Passwords do not match';
           }
         } 
-        if ($_FILES['profile_image']['error'] == 4) {
-          $data['profile_err'] = 'Upload a image';
-     
-        }
+      
+      
 
-        if(empty($data['email_err']) && empty($data['name_err']) && empty($data['password_err']) && empty($data['confirm_password_err']) && empty($data['contact_no_err']) && empty($data['nic_err']) && empty($data['address_err']) && empty($data['dob_err'])){
-          if ($_FILES['profile_image']['error'] == 4) {
-            $data['profile_err'] = 'Upload a image';
-        } else {
-            if (uploadImage($_FILES['profile_image']['tmp_name'], $data['profile_image_name'], '/img/img_upload/Admin/')) {
-              $data['profile_err'] = '';
-  
-            } else {
-                $data['profile_err'] = 'Error uploading the profile image';
-            }
-        }
-        }
-
-        if(empty($data['email_err']) && empty($data['name_err']) && empty($data['password_err']) && empty($data['confirm_password_err']) && empty($data['contact_no_err']) && empty($data['nic_err']) && empty($data['address_err']) && empty($data['dob_err']) && empty($data['profile_err'])){
+        if(empty($data['email_err']) && empty($data['name_err']) && empty($data['password_err']) && empty($data['confirm_password_err']) && empty($data['contact_no_err']) && empty($data['nic_err']) && empty($data['address_err']) && empty($data['dob_err']) ){
           // Validated
           $data['password'] = password_hash($data['password'], PASSWORD_DEFAULT);
           if($this->adminModel->register_admin($data)){
@@ -1705,7 +1794,13 @@
       $admin_by_id = $this->adminModel->getAdminByID($id);
       $this->adminModel->admin_delete($id);
       $admin = $this->adminModel->get_all();
-      deleteImage("C:\\xampp\\htdocs\\ecoplus\\public\\img\\img_upload\\Admin\\" . $admin_by_id->image);
+     
+      if($admin_by_id->image=="profile.png"){
+
+      }else{
+        deleteImage("C:\\xampp\\htdocs\\ecoplus\\public\\img\\img_upload\\Admin\\" . $admin_by_id->image);
+      }
+      
       $data = [
         'admin' => $admin,
         'confirm_delete' =>'',
